fix inconsistencies

+++ Sources/App/Models/API/AnvilAnalyzeProfilePreviewResponse.swift
@@ -16,7 +16,7 @@ struct AnvilAnalyzeProfilePreviewDebugDTO: Content, Sendable, Equatable {
     let centroid: StormSetupCentroid
     let selectedMessageCount: Int
     let selectedPressureLevels: [Int]
-    let surfacePressureMb: Int
+    let surfacePressureMb: Double
     let surfaceSubsetCacheHit: Bool
     let rangeCount: Int
     let totalSelectedRangeBytes: Int64
